refactor(chat): extract chat_manager/deps_offer.rs (#292)
Task 7 of the split plan: OfferDeps and its WorktreeOfferDeps impl
move into their own file, owning their own construction via a
build() helper the constructor now calls.
